Just a hint to better wxThread deletion.
Still don't think I've got it right, so have left some commented out hints as
to proper deleting. Docs say that the threads "Wait" or "Destroy" functions
should not be called in anthers destructor but on a OnClose(wxCloseEvent&)
event function. When I tried this the app closed down but there were memory
leaks.