Add ability to transfer profile to device
The profile will continue to be sent until all of it has been transferred. As
data starts getting read in from the device it is immediately sent to the
display to be processed. Changes to the biases or device settings are
transferred between profiles. New profiles will only start to be transferred
once the old profiles are done.