Add profile rap-around code to TransferData function
When the end of the profile is about to be transferred to the device the
command queue is checked to see if there is a need to stop transferring data in
order to process commands. If there are commands to be processed then only the
last of the current profile is transferred thus ensuring that the device is
left at a safe (holding) state. All data is then read from the device that
corresponds to the current settings and profile. The TransferData function is
then returned and the rest of the command queue is then processed.

However, if there is no commands that need processing, data from the start of
the profile is added to the end of the current profile (rapped round) and the
transfer continues.

Note that measurement on the device is paused (as there is no profile) while
the command queue is being processed. An argument can be made that since the
commands will be changing bias values or other data measurement values that no
valid measurements can be made during this time. Thus pausing measurement on
the device while the device is being held at a safe state is an acceptable
response.