Add circular buffer template
Each new element is added to the buffer by pushing it to the head of the
buffer. When the data needs to be read an array pointer to the data is returned
from the GetData function. The earliest data can is found in at the zeroth
index with the most resent is at higher indexes.